extract contexts into its own file
