+++ packages/pwa/app/hooks/use-product-view-modal.js
@@ -8,7 +8,8 @@
 import {useEffect, useState} from 'react'
 import {rebuildPathWithParams, removeQueryParamsFromPath} from '../utils/url'
 import {useHistory, useLocation} from 'react-router-dom'
-import {isError, useCommerceAPI} from '../commerce-api/utils'
+import {useCommerceAPI} from '../commerce-api/contexts'
+import {isError} from '../commerce-api/utils'
 import {useVariant} from './use-variant'
 import {useToast} from './use-toast'
 import {useIntl} from 'react-intl'
